Remove module namespace usage
