+++ pages/boac/boac_admin_page.rb
@@ -1,0 +1,15 @@
+require_relative '../../util/spec_helper'
+
+class BOACAdminPage
+
+  include PageObject
+  include Logging
+  include Page
+  include BOACPages
+
+  # Loads the admin page
+  def load_page
+    navigate_to "#{BOACUtils.base_url}/admin"
+  end
+
+end
