+++ pages/boac/boac_api_admin_page.rb
@@ -1,6 +1,6 @@
 require_relative '../../util/spec_helper'
 
-class ApiAdminPage
+class BOACApiAdminPage
 
   include PageObject
   include Logging
