+++ pages/boac/boac_api_user_analytics_page.rb
@@ -1,6 +1,6 @@
 require_relative '../../util/spec_helper'
 
-class ApiUserAnalyticsPage
+class BOACApiUserAnalyticsPage
 
   include PageObject
   include Logging
