+++ util/spec_helper.rb
@@ -72,21 +72,24 @@ require_relative '../pages/junction/canvas_mailing_lists_page'
 require_relative '../pages/junction/canvas_mailing_list_page'
 require_relative '../pages/junction/canvas_e_grades_export_page'
 require_relative '../pages/boac/boac_pages'
-require_relative '../pages/boac/user_list_pages'
-require_relative '../pages/boac/admin_page'
-require_relative '../pages/boac/home_page'
-require_relative '../pages/boac/cohort_pages'
-require_relative '../pages/boac/class_pages'
-require_relative '../pages/boac/class_list_view_page'
-require_relative '../pages/boac/class_matrix_view_page'
-require_relative '../pages/boac/curated_group_page'
-require_relative '../pages/boac/filtered_cohort_page'
-require_relative '../pages/boac/search_results_page'
-require_relative '../pages/boac/student_page'
-require_relative '../pages/boac/teams_list_page'
-require_relative '../pages/boac/api_admin_page'
-require_relative '../pages/boac/api_section_page'
-require_relative '../pages/boac/api_user_analytics_page'
+require_relative '../pages/boac/boac_add_curated_modal_pages'
+require_relative '../pages/boac/boac_add_curated_selector_pages'
+require_relative '../pages/boac/boac_list_view_pages'
+require_relative '../pages/boac/boac_user_list_pages'
+require_relative '../pages/boac/boac_admin_page'
+require_relative '../pages/boac/boac_home_page'
+require_relative '../pages/boac/boac_cohort_pages'
+require_relative '../pages/boac/boac_class_pages'
+require_relative '../pages/boac/boac_class_list_view_page'
+require_relative '../pages/boac/boac_class_matrix_view_page'
+require_relative '../pages/boac/boac_curated_group_page'
+require_relative '../pages/boac/boac_filtered_cohort_page'
+require_relative '../pages/boac/boac_search_results_page'
+require_relative '../pages/boac/boac_student_page'
+require_relative '../pages/boac/boac_teams_list_page'
+require_relative '../pages/boac/boac_api_admin_page'
+require_relative '../pages/boac/boac_api_section_page'
+require_relative '../pages/boac/boac_api_user_analytics_page'
 require_relative '../pages/salesforce_page'
 require_relative '../pages/suitec/suite_c_pages'
 require_relative '../pages/suitec/asset_library_list_view_page'
